QID renamed and moved to networking - builds lib-tempo
