+++ src/lib-tempo/include/tempo/component/ComponentPlayerRemoteServer.hpp
@@ -6,7 +6,7 @@
 
 #include <tempo/component/ComponentGridMotion.hpp>
 #include <tempo/system/SystemID.hpp>
-#include <tempo/entity/SystemQID.hpp>
+#include <tempo/network/QueueID.hpp>
 #include <tempo/network/queue.hpp>
 #include <tempo/network/server.hpp>
 #include <tempo/time.hpp>
