+++ src/lib-tempo/include/tempo/network/base.hpp
@@ -1,10 +1,8 @@
 #ifndef TEMPO_NETWORK_BASE_HPP
 #define TEMPO_NETWORK_BASE_HPP
 
-#include <tempo/entity/SystemQID.hpp>
-#include <tempo/entity/SystemQID.hpp>
 #include <tempo/network/queue.hpp>
-#include <tempo/network/queue.hpp>
+#include <tempo/network/QueueID.hpp>
 #include <tempo/structures.hpp>
 
 #include <SFML/Network.hpp>
