+++ src/lib-tempo/include/tempo/system/SystemPlayerRemoteServer.hpp
@@ -7,8 +7,8 @@
 #include <tempo/component/ComponentGridMotion.hpp>
 #include <tempo/component/ComponentPlayerRemoteServer.hpp>
 #include <tempo/system/SystemID.hpp>
-#include <tempo/entity/SystemQID.hpp>
 #include <tempo/network/queue.hpp>
+#include <tempo/network/QueueID.hpp>
 #include <tempo/network/server.hpp>
 #include <tempo/time.hpp>
